refactor: split test_sa.py into focused test files with factory_boy
Split the monolithic test_sa.py (1003 lines) into:
- tests/ext/sa/conftest.py — models, fixtures, LoanRecordFactory,
  SettlementRecordFactory
- tests/ext/sa/test_sa_money_type.py — MoneyType round-trips + None
- tests/ext/sa/test_sa_rate_type.py — RateType + InterestRateType
- tests/ext/sa/test_sa_bridge.py — bridge decorators, balance, balance_at
- tests/ext/sa/test_sa_integration.py — real Loan engine round-trip

Add factory-boy dev dependency. Factories eliminate verbose model
construction and avoid hardcoded IDs across tests.

Made-with: Cursor
