+++ pyproject.toml
@@ -44,6 +44,7 @@ tox = "^3.25.1"
 hypothesis = {version = "^6.151.9", python = ">=3.10"}
 marshmallow = "^3.20.0"
 sqlalchemy = "^2.0"
+factory-boy = "^3.3.3"
 
 [tool.poetry.group.docs.dependencies]
 mkdocs = "^1.6.1"
